# docs/swagger/movies.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Movies API",
    "version": "1.0",
    "contact": {
      "name": "Lovelyo",
      "url": "https://lovelyoyrmia.com",
      "email": "lovelyoyrmia@gmail.com"
    }
  },
  "tags": [
    {
      "name": "MovieServices"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/movies": {
      "get": {
        "summary": "List Movies",
        "description": "Use this API to get list of movies",
        "operationId": "MovieServices_ListMovies",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListMoviesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "MovieServices"
        ]
      },
      "post": {
        "summary": "Add Movie",
        "description": "Use this API to add a new movie",
        "operationId": "MovieServices_AddMovie",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbMovieResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbMovieParams"
            }
          }
        ],
        "tags": [
          "MovieServices"
        ]
      }
    },
    "/movies/{id}": {
      "get": {
        "summary": "Detail Movie",
        "description": "Use this API to get detail of movie",
        "operationId": "MovieServices_DetailMovie",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbMovieResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "MovieServices"
        ]
      },
      "delete": {
        "summary": "Delete Movie",
        "description": "Use this API to delete movie",
        "operationId": "MovieServices_DeleteMovie",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbMovieResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "MovieServices"
        ]
      },
      "patch": {
        "summary": "Update Movie",
        "description": "Use this API to update movie",
        "operationId": "MovieServices_UpdateMovie",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbMovieResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MovieServicesUpdateMovieBody"
            }
          }
        ],
        "tags": [
          "MovieServices"
        ]
      }
    }
  },
  "definitions": {
    "MovieServicesUpdateMovieBody": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "rating": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "pbListMoviesResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbMovie"
          }
        }
      }
    },
    "pbMovie": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "rating": {
          "type": "number",
          "format": "float"
        },
        "createdAt": {
          "type": "string"
        },
        "updateAt": {
          "type": "string"
        }
      }
    },
    "pbMovieParams": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "rating": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "pbMovieResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/pbMovie"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}